--- example/src/main/kotlin/com/klee0kai/example/engine/CppCallbacks.kt
@@ -1,13 +1,0 @@
-package com.klee0kai.example.engine
-
-import com.github.klee0kai.brooklyn.JniMirror
-import com.klee0kai.example.model.Simple
-
-@JniMirror
-object CppCallbacks {
-
-    fun receiveSimple(simple: Simple){
-
-    }
-
-}
