--- example/src/main/kotlin/com/klee0kai/example/mirrors/JniCounter.kt
@@ -1,14 +1,0 @@
-package com.klee0kai.example.mirrors
-
-import com.github.klee0kai.brooklyn.JniMirror
-
-@JniMirror
-object JniCounter {
-
-    var count: Int = 0
-
-    fun inc() {
-        count++
-    }
-
-}
