+++ tests/src/main/kotlin/com/klee0kai/tests/mirrors/JniCounter.kt
@@ -1,0 +1,14 @@
+package com.klee0kai.tests.mirrors
+
+import com.github.klee0kai.brooklyn.JniMirror
+
+@JniMirror
+object JniCounter {
+
+    var count: Int = 0
+
+    fun inc() {
+        count++
+    }
+
+}
